The Demo Update
This is as ready as I alone will get it before our presentation:
SettingsActivity is now "RulesActivity" - Displays 4 rules of the game.
For the sake of demos, the first time a game is inserted, it allows the
creator to play another turn.
Matt got a game_view layout working so that we have a nice list view for
ViewGamesActivity.

Some things still to check:
Friends list allows multiple additions of friends (can cause trouble
when trying to insert a friend that already exists)
I think friends with spaces in their names may cause some db issues, not
certain tho...
And of course, things are gonna get crazier when we start implementing
server stuff