Just Kidding, This is the Demo Update.0.2
Made it so that deleting a friend also deletes any active games they may
have with the player.  And also changed some things in Friend and Game
ContentProviders:
A couple spacing things
I now see what you mean by needing to have quotes for some db
statements.  You added them before running the query in the
ContentProvider, but I was doing it within the method in the
ContentProvider.